Flash-card-Challenge
Python script that takes a file containing flash card questions and answers
as an argument and quizzes the user based on the contents of that file until
the user quits the program. Questions should be selected randomly (as opposed
to going in order through the file), and the user should type in their guess.
The script should say whether or not a guess is correct and provide the
correct answer if an incorrect answer is given.